+++ src/locales/en/search.js
@@ -5,5 +5,7 @@ export default {
   searchPlaceholder: 'Search',
   searchResult: 'result',
   searchResults: 'results',
+  searchLoading: 'The index is still loading',
+  searchError: 'Unable to load the index',
   searchNotFound: 'Nothing found',
 };