+++ src/locales/fi/search.js
@@ -2,9 +2,10 @@
 // Licensed under the MIT License
 
 export default {
-  searchPagesTitle: 'Sivut',
   searchPlaceholder: 'Haku',
   searchResult: 'tulos',
   searchResults: 'tulosta',
+  searchLoading: 'Hakemistoa ladataan',
+  searchError: 'Hakemiston lataaminen epäonnistui',
   searchNotFound: 'Ei tuloksia',
 };